API Documentation

Thermodynamics

Breeze.Thermodynamics.AtmosphereThermodynamics — Type

AtmosphereThermodynamics(FT = Oceananigans.defaults.FloatType;
                         gravitational_acceleration = 9.81,
                         molar_gas_constant = 8.314462618,
                         dry_air_molar_mass = 0.02897,
                         dry_air_heat_capacity = 1005,
                         vapor_molar_mass = 0.018015,
                         vapor_heat_capacity = 1850,
                         saturation = Saturation(FT),
                         condensation = water_condensation(FT),
                         deposition = water_deposition(FT))

Create AtmosphereThermodynamics with parameters that correpsond to the composition of dry air in Earth's atmosphere and water vapor. The default isnothing(condensation) implies thermodynamics appropriate for unsaturated and therefore non-condensing air.

MoistAirBuoyancies

Breeze.MoistAirBuoyancies.MoistAirBuoyancy — Type

MoistAirBuoyancy(FT=Oceananigans.defaults.FloatType;
                 thermodynamics = AtmosphereThermodynamics(FT),
                 reference_constants = ReferenceConstants{FT}(101325, 290))

Return a MoistAirBuoyancy.

AtmosphereModels

Breeze.AtmosphereModels.AtmosphereModel — Method

AtmosphereModel(grid;
                clock = Clock(grid),
                thermodynamics = AtmosphereThermodynamics(eltype(grid)),
                formulation = default_formulation(grid, thermodynamics),
                absolute_humidity = DefaultValue(),
                tracers = tuple(),
                coriolis = nothing,
                boundary_conditions = NamedTuple(),
                forcing = NamedTuple(),
                advection = WENO(order=5),
                microphysics = WarmPhaseSaturationAdjustment(),
                timestepper = :RungeKutta3)

Return an AtmosphereModel that uses the anelastic approximation following Pauluis (2008).